Abstract AddClientCommandParserTest and add AddPastryCommandParserTest
